replace remaining icons
