+++ packages/client/lib/components/ReportForm/ReportForm.jsx
@@ -35,9 +35,7 @@ const ReportForm = () => {
     <div className="bg-background">
       <form
         noValidate
-        className={`flex flex-col h-screen overflow-hidden ${
-          step !== 5 ? 'p-2' : ''
-        }`}
+        className={`flex flex-col h-screen ${step !== 5 ? 'p-2' : ''}`}
       >
         {!isStepFive && (
           <h1 className="text-2xl font-medium">Report your incident</h1>
