+++ packages/client/lib/components/common/icons/SVGIcon.jsx
@@ -1,9 +1,12 @@
 import {
   ArrowUturnLeftIcon,
+  ArrowUturnRightIcon,
+  CheckIcon,
   DocumentMagnifyingGlassIcon,
   HomeIcon,
   MagnifyingGlassIcon,
   MapPinIcon,
+  NoSymbolIcon,
   PhotoIcon,
   PlusCircleIcon,
   UserIcon,
@@ -11,36 +14,33 @@ import {
   XMarkIcon,
 } from '@heroicons/react/24/outline';
 
-export const XMark = ({ handleClick }) => {
-  return (
-    <button
-      onClick={(e) => {
-        e.preventDefault();
-        handleClick();
-      }}
-    >
-      <XCircleIcon className="z-10 absolute top-0.5 right-0.5 text-foreground stroke-2 h-6 w-6" />
-    </button>
-  );
-};
-
-// export const Icon = () => {
-//     return (
-//         <CheckIcon className="w-6 h-6 pl-1 text-background" />
-//   )
-// }
+// export const XMark = ({ handleClick }) => {
+//   return (
+//     <button
+//       onClick={(e) => {
+//         e.preventDefault();
+//         handleClick();
+//       }}
+//     >
+//       <XCircleIcon className="z-10 absolute top-0.5 right-0.5 text-foreground stroke-2 h-6 w-6" />
+//     </button>
+//   );
+// };
 
 /**
  * map of icon names to import paths
  */
 const iconMap = new Map()
   .set('arrowUturnLeftIcon', ArrowUturnLeftIcon)
+  .set('arrowUturnRightIcon', ArrowUturnRightIcon)
+  .set('checkIcon', CheckIcon)
   .set('documentMagnifyingGlassIcon', DocumentMagnifyingGlassIcon)
   .set('homeIcon', HomeIcon)
   .set('magnifyingGlassIcon', MagnifyingGlassIcon)
   .set('mapPinIcon', MapPinIcon)
   .set('plusCircleIcon', PlusCircleIcon)
   .set('photoIcon', PhotoIcon)
+  .set('noSymbolIcon', NoSymbolIcon)
   .set('userIcon', UserIcon)
   .set('xCircleIcon', XCircleIcon)
   .set('xMarkIcon', XMarkIcon);
